+++ changelog.md
@@ -9,6 +9,7 @@ __0.9.6__
 * fix `container` meta parameter in `compositefilter` after each `apply` call (could be overwritten deeper in apply chain)
 * canny gradient pruning in haar detector computed using wrong indices in previous optimisations, fixed
 * the way extra filter inputs were handled has a bug, if same image is used as extra input in more than one filter and image is updated through another filter, it is possible depending on order of application that some filters will get the previous version of the image as input (because it is cached and not resent to save bandwidth) while only the first filter will get the updated (correct) version, fixed
+* add some machine learning algorithms, `kmeans`, `kmedoids`, move `connected_components` algorithm under `machinelearning` package
 * update references, examples